+++ docs/News.rst
@@ -14,6 +14,8 @@ Build
 * Change URLs for ``oursql`` in ``extras_require`` in ``setup.py``.
   Provide separate URLs for Python 2.7 and 3.4+.
 
+* Add ``mariadb`` in ``extras_require`` in ``setup.py``.
+
 SQLObject 3.9.0
 ===============